+++ udata_search_service/infrastructure/consumers.py
@@ -24,6 +24,11 @@ class DatasetConsumer(Dataset):
         data["organization_name"] = organization.get('name') if organization else None
         data["organization_badges"] = organization.get('badges') if organization else None
 
+        resources = data["resources"]
+        data["resources_ids"] = [res.get("id") for res in resources]
+        data["resources_titles"] = [res.get("title") for res in resources]
+        del data["resources"]
+
         data["concat_title_org"] = get_concat_title_org(data["title"], data['acronym'], data['organization_name'])
         data["geozones"] = [zone.get("id") for zone in data.get("geozones", [])]