+++ udata_search_service/infrastructure/search_clients.py
@@ -143,6 +143,8 @@ class SearchableDataset(IndexDocument):
     reuses = Float()
     featured = Integer()
     resources_count = Integer()
+    resources_ids = Keyword(multi=True)
+    resources_titles = Text(analyzer=dgv_analyzer)
     concat_title_org = Text(analyzer=dgv_analyzer)
     temporal_coverage_start = Date()
     temporal_coverage_end = Date()
@@ -220,7 +222,7 @@ class ElasticClient:
             search = search.query('bool', should=[
                     query.Q(
                         'function_score',
-                        query=query.Bool(should=[query.MultiMatch(query=query_text, type='phrase', fields=['name^15', 'acronym^15', 'description^8'])]),
+                        query=query.Bool(should=[query.MultiMatch(query=query_text, type='phrase', fields=['id^15', 'name^15', 'acronym^15', 'description^8'])]),
                         functions=organizations_score_functions
                     ),
                     query.Q(
@@ -228,7 +230,7 @@ class ElasticClient:
                         query=query.Bool(should=[query.MultiMatch(
                             query=query_text,
                             type='cross_fields',
-                            fields=['name^7', 'acronym^7', 'description^4'],
+                            fields=['id^15', 'name^7', 'acronym^7', 'description^4'],
                             operator="and")]),
                         functions=organizations_score_functions
                     ),
@@ -283,7 +285,11 @@ class ElasticClient:
                 should=[
                     query.Q(
                         'function_score',
-                        query=query.Bool(should=[query.MultiMatch(query=query_text, type='phrase', fields=['title^15', 'acronym^15', 'description^8', 'organization_name^8'])]),
+                        query=query.Bool(should=[query.MultiMatch(
+                            query=query_text,
+                            type='phrase',
+                            fields=['id^15', 'title^15', 'acronym^15', 'description^8', 'organization_name^8', 'resources_ids^8', 'resources_titles^5']
+                        )]),
                         functions=datasets_score_functions
                     ),
                     query.Q(
@@ -296,7 +302,7 @@ class ElasticClient:
                         query=query.Bool(should=[query.MultiMatch(
                             query=query_text,
                             type='cross_fields',
-                            fields=['title^7', 'acronym^7', 'description^4', 'organization_name^4'],
+                            fields=['id^7', 'title^7', 'acronym^7', 'description^4', 'organization_name^4', 'resources_ids^4', 'resources_titles^2'],
                             operator="and")]),
                         functions=datasets_score_functions
                     ),
@@ -338,7 +344,7 @@ class ElasticClient:
             search = search.query('bool', should=[
                     query.Q(
                         'function_score',
-                        query=query.Bool(should=[query.MultiMatch(query=query_text, type='phrase', fields=['title^15', 'description^8', 'organization_name^8'])]),
+                        query=query.Bool(should=[query.MultiMatch(query=query_text, type='phrase', fields=['id^15', 'title^15', 'description^8', 'organization_name^8'])]),
                         functions=reuses_score_functions
                     ),
                     query.Q(
@@ -346,7 +352,7 @@ class ElasticClient:
                         query=query.Bool(should=[query.MultiMatch(
                             query=query_text,
                             type='cross_fields',
-                            fields=['title^7', 'description^4', 'organization_name^4'],
+                            fields=['id^7', 'title^7', 'description^4', 'organization_name^4'],
                             operator="and")]),
                         functions=reuses_score_functions
                     ),
@@ -387,7 +393,7 @@ class ElasticClient:
             search = search.query('bool', should=[
                     query.Q(
                         'function_score',
-                        query=query.Bool(should=[query.MultiMatch(query=query_text, type='phrase', fields=['title^15', 'description^8', 'organization_name^8'])]),
+                        query=query.Bool(should=[query.MultiMatch(query=query_text, type='phrase', fields=['id^15', 'title^15', 'description^8', 'organization_name^8'])]),
                         functions=dataservices_score_functions
                     ),
                     query.Q(
@@ -395,7 +401,7 @@ class ElasticClient:
                         query=query.Bool(should=[query.MultiMatch(
                             query=query_text,
                             type='cross_fields',
-                            fields=['title^7', 'description^4', 'organization_name^4'],
+                            fields=['id^7', 'title^7', 'description^4', 'organization_name^4'],
                             operator="and")]),
                         functions=dataservices_score_functions
                     ),